refactor(test-config): extract CommonTestSchedulerConfig and centralize task scheduler beans
- Removed duplicated task scheduler beans from IntegrationTestConfig and ServiceTestConfig.
- Introduced CommonTestSchedulerConfig and imported it into relevant configurations.
