+++ backend/app/src/test/java/coffeeshout/global/config/ServiceTestConfig.java
@@ -1,14 +1,13 @@
 package coffeeshout.global.config;
 
 import coffeeshout.gamecommon.flow.FlowScheduler;
-import coffeeshout.support.ShutDownTestScheduler;
 import coffeeshout.support.TestTaskScheduler;
 import java.time.Clock;
-import org.mockito.Answers;
 import org.mockito.Mockito;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.Primary;
 import org.springframework.context.annotation.Profile;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
@@ -16,13 +15,8 @@ import org.springframework.scheduling.TaskScheduler;
 
 @TestConfiguration(proxyBeanMethods = false)
 @Profile("test")
+@Import(CommonTestSchedulerConfig.class)
 public class ServiceTestConfig {
-
-    @Bean(name = "taskScheduler")
-    @Primary
-    public TaskScheduler noOpTaskScheduler() {
-        return Mockito.mock(TaskScheduler.class, Answers.RETURNS_MOCKS);
-    }
 
     @Bean(name = "cardGameFlowScheduler")
     @Primary
@@ -38,11 +32,6 @@ public class ServiceTestConfig {
     @Bean(name = "ladderFlowScheduler")
     public FlowScheduler mockLadderFlowScheduler() {
         return Mockito.mock(FlowScheduler.class);
-    }
-
-    @Bean(name = "delayRemovalScheduler")
-    public TaskScheduler testDelayRemovalScheduler() {
-        return new ShutDownTestScheduler();
     }
 
     @Bean(name = "racingGameScheduler")
